Generate a title image for each post
